Cleaning the function of processing extrinsics
